+++ src/pdfview-ng.android.ts
@@ -3,17 +3,50 @@
 import {
   PDFViewNgCommon,
   srcProperty,
-  defaultpageProperty
+  defaultpageProperty,
+  BookmarkCommon,
+  bookmarkPathProperty,
+  bookmarkLabelProperty
 } from "./pdfview-ng.common";
 import * as fs from "tns-core-modules/file-system";
 import pdfviewer = com.github.barteksc.pdfviewer;
 import * as http from "tns-core-modules/http";
+
+export class Bookmark extends BookmarkCommon {
+  private nativeitem: pdfviewer.Bookmark;
+
+  public get android(): pdfviewer.Bookmark {
+    return this.nativeitem;
+  }
+
+  public set android(value: pdfviewer.Bookmark) {
+    this.nativeitem = value;
+  }
+
+  public getTitle(): string {
+    return this.nativeitem.getTitle();
+  }
+
+  public getChildren(): Bookmark[] {
+    let list: Bookmark[] = [];
+    let nativechildren = this.android.getChildren();
+    for (let i=0;i<nativechildren.size();i++){
+      let item = nativechildren.get(i);
+      let boxed = new Bookmark();
+      boxed.android = item;
+      list.push(boxed);
+    }
+    return list;
+  }
+}
 
 export class PDFViewNg extends PDFViewNgCommon {
   private promise: Promise<void>;
   private tempFolder = fs.knownFolders.temp().getFolder("PDFViewer.temp/");
   private value_src: string;
   private value_default_page: string = "0";
+  private value_bookmark_path: number[];
+  private value_bookmark_label: string;
 
   public get android() {
     return this.nativeView as pdfviewer.PDFView;
@@ -29,11 +62,20 @@ export class PDFViewNg extends PDFViewNgCommon {
   }
 
   public [srcProperty.setNative](value: string) {
+    console.error("MUUUH",value);
     this.value_src = value;
   }
 
   public [defaultpageProperty.setNative](value: string) {
     this.value_default_page = value;
+  }
+
+  public [bookmarkPathProperty.setNative](value: string) {
+    this.value_bookmark_path = value.split(",").map(item=>parseInt(item));
+  }
+
+  public [bookmarkLabelProperty.setNative](value: string) {
+    this.value_bookmark_label = value;
   }
 
   public onLoaded(): void {
@@ -42,6 +84,29 @@ export class PDFViewNg extends PDFViewNgCommon {
   }
 
   public loadPDF(src: string, default_page: number) : Promise<any> {
+    let that = this;
+    let onLoadHandler = (() => {
+      const pdfViewRef = new WeakRef(this);
+  
+      return new pdfviewer.listener.OnLoadCompleteListener({
+        loadComplete: numPages => {
+          console.log("PDFViewNg Android (Step 5) Pages loaded: " + numPages);
+          
+          setTimeout(()=>{
+            if (that.value_bookmark_label) {
+              console.log("PDFViewNg Android (Step 6) Go to outlinelabel: ", src, that.value_bookmark_label);
+              that.goToBookmarkByLabel(""+that.value_bookmark_label);
+            }else if (that.value_bookmark_path){
+              console.log("PDFViewNg Android (Step 6) Go to outlinepath: ", src, that.value_bookmark_path);
+              that.goToBookmarkByPath(that.value_bookmark_path);
+            }
+    
+            PDFViewNgCommon.notifyOfEvent(PDFViewNgCommon.loadEvent, pdfViewRef);
+          },1);
+        }
+      });
+    })();
+
     if (!src || !this.android) {
       return Promise.reject('no parameters');
     }
@@ -75,20 +140,46 @@ export class PDFViewNg extends PDFViewNgCommon {
 
         this.android
           .fromUri(uri)
-          .onLoad(this.onLoadHandler)
+          .onLoad(onLoadHandler)
           .defaultPage(default_page)
           .load();
       });
   }
 
-  private onLoadHandler = (() => {
-    const pdfViewRef = new WeakRef(this);
+  public getPageCount(): number {
+    return this.android.getPageCount();
+  }
 
-    return new pdfviewer.listener.OnLoadCompleteListener({
-      loadComplete: numPages => {
-        console.log("PDFViewNg Android (Step 5) Pages loaded: " + numPages);
-        PDFViewNgCommon.notifyOfEvent(PDFViewNgCommon.loadEvent, pdfViewRef);
-      }
-    });
-  })();
+  public goToPage(index: number) : void {
+    this.android.jumpTo(index, false);
+  }
+
+  public goToFirstPage() : void {
+    this.android.jumpTo(0, false);
+  }
+
+  public goToLastPage() : void {
+    this.android.jumpTo(this.android.getPageCount() - 1, false);
+  }
+
+  public goToBookmark(bookmark: Bookmark): void {
+    this.android.jumpTo(bookmark.android.getPageIdx(), false);
+  }
+
+  public getAuthor(): string {
+    let meta = this.android.getDocumentMeta();
+    return meta.getAuthor();
+  }
+
+  public getBookmarks(): Bookmark[] {
+    let list: Bookmark[] = [];
+    let tableOfContents = this.android.getTableOfContents();
+    for (let i=0;i<tableOfContents.size();i++){
+      let item: pdfviewer.Bookmark = tableOfContents.get(i);
+      let boxed = new Bookmark();
+      boxed.android = item;
+      list.push(boxed);
+    }
+    return list;
+  }
 }